+++ clinicadl/commandline/pipelines/interpret/options.py
@@ -1,8 +1,8 @@
 import click
 
-from clinicadl.config.config.pipelines.interpret import InterpretConfig
 from clinicadl.config.config_utils import get_default_from_config_class as get_default
 from clinicadl.config.config_utils import get_type_from_config_class as get_type
+from clinicadl.interpret.config import InterpretConfig
 
 # interpret specific
 name = click.argument(
