+++ clinicadl/commandline/pipelines/predict/options.py
@@ -1,8 +1,8 @@
 import click
 
-from clinicadl.config.config.pipelines.predict import PredictConfig
 from clinicadl.config.config_utils import get_default_from_config_class as get_default
 from clinicadl.config.config_utils import get_type_from_config_class as get_type
+from clinicadl.predict.config import PredictConfig
 
 # predict specific
 use_labels = click.option(
