+++ clinicadl/commandline/pipelines/train/classification/cli.py
@@ -23,8 +23,8 @@ from clinicadl.commandline.pipelines.train.classification import (
 from clinicadl.commandline.pipelines.transfer_learning import (
     options as transfer_learning,
 )
-from clinicadl.config.config.pipelines.task.classification import ClassificationConfig
 from clinicadl.train.utils import merge_cli_and_config_file_options
+from clinicadl.trainer.config.classification import ClassificationConfig
 from clinicadl.trainer.trainer import Trainer
 from clinicadl.utils.enum import Task
