+++ clinicadl/commandline/pipelines/train/reconstruction/options.py
@@ -1,11 +1,11 @@
 import click
 
-from clinicadl.config.config.pipelines.task.reconstruction import (
+from clinicadl.config.config_utils import get_default_from_config_class as get_default
+from clinicadl.config.config_utils import get_type_from_config_class as get_type
+from clinicadl.trainer.config.reconstruction import (
     NetworkConfig,
     ValidationConfig,
 )
-from clinicadl.config.config_utils import get_default_from_config_class as get_default
-from clinicadl.config.config_utils import get_type_from_config_class as get_type
 
 # Model
 architecture = click.option(
