+++ clinicadl/commandline/pipelines/train/regression/cli.py
@@ -21,8 +21,8 @@ from clinicadl.commandline.pipelines.train.regression import options as regressi
 from clinicadl.commandline.pipelines.transfer_learning import (
     options as transfer_learning,
 )
-from clinicadl.config.config.pipelines.task.regression import RegressionConfig
 from clinicadl.train.utils import merge_cli_and_config_file_options
+from clinicadl.trainer.config.regression import RegressionConfig
 from clinicadl.trainer.trainer import Trainer
 from clinicadl.utils.enum import Task
