+++ clinicadl/commandline/pipelines/train/regression/options.py
@@ -1,12 +1,12 @@
 import click
 
-from clinicadl.config.config.pipelines.task.regression import (
+from clinicadl.config.config_utils import get_default_from_config_class as get_default
+from clinicadl.config.config_utils import get_type_from_config_class as get_type
+from clinicadl.trainer.config.regression import (
     DataConfig,
     NetworkConfig,
     ValidationConfig,
 )
-from clinicadl.config.config_utils import get_default_from_config_class as get_default
-from clinicadl.config.config_utils import get_type_from_config_class as get_type
 
 # Data
 label = click.option(
