+++ clinicadl/random_search/random_search_config.py
@@ -5,13 +5,13 @@ from typing import TYPE_CHECKING, Any, Dict, Optional, Tuple, Type, Union
 
 from pydantic import BaseModel, ConfigDict, PositiveInt, field_validator
 
-from clinicadl.config.config.pipelines.task.classification import (
+from clinicadl.config.config_utils import get_type_from_config_class as get_type
+from clinicadl.trainer.config.classification import (
     ClassificationConfig as BaseClassificationConfig,
 )
-from clinicadl.config.config.pipelines.task.regression import (
+from clinicadl.trainer.config.regression import (
     RegressionConfig as BaseRegressionConfig,
 )
-from clinicadl.config.config_utils import get_type_from_config_class as get_type
 from clinicadl.utils.enum import Normalization, Pooling, Task
 
 if TYPE_CHECKING:
