+++ clinicadl/train/tasks_utils.py
@@ -1,7 +1,7 @@
 # TODO: to put in trainer ? trainer_utils.py ?
 from typing import Type, Union
 
-from clinicadl.config.config.pipelines.train import TrainConfig
+from clinicadl.trainer.config.train import TrainConfig
 from clinicadl.utils.enum import Task
 
 
@@ -16,15 +16,15 @@ def create_training_config(task: Union[str, Task]) -> Type[TrainConfig]:
     """
     task = Task(task)
     if task == Task.CLASSIFICATION:
-        from clinicadl.config.config.pipelines.task.classification import (
+        from clinicadl.trainer.config.classification import (
             ClassificationConfig as Config,
         )
     elif task == Task.REGRESSION:
-        from clinicadl.config.config.pipelines.task.regression import (
+        from clinicadl.trainer.config.regression import (
             RegressionConfig as Config,
         )
     elif task == Task.RECONSTRUCTION:
-        from clinicadl.config.config.pipelines.task.reconstruction import (
+        from clinicadl.trainer.config.reconstruction import (
             ReconstructionConfig as Config,
         )
     return Config
