+++ clinicadl/trainer/config/classification.py
@@ -6,7 +6,7 @@ from pydantic import computed_field, field_validator
 from clinicadl.caps_dataset.data_config import DataConfig as BaseDataConfig
 from clinicadl.config.config import NetworkConfig as BaseNetworkConfig
 from clinicadl.config.config import ValidationConfig as BaseValidationConfig
-from clinicadl.config.config.pipelines.train import TrainConfig
+from clinicadl.trainer.config.train import TrainConfig
 from clinicadl.utils.enum import ClassificationLoss, ClassificationMetric, Task
 
 logger = getLogger("clinicadl.classification_config")
